+++ src/test/runTest.ts
@@ -3,24 +3,18 @@ import { runTests } from '@vscode/test-electron';
 
 async function main() {
     try {
+        // The folder containing the Extension Manifest package.json
+        // Passed to `--extensionDevelopmentPath`
         const extensionDevelopmentPath = path.resolve(__dirname, '../../');
-        console.log('Extension development path:', extensionDevelopmentPath);
 
+        // The path to the extension test script
+        // Passed to --extensionTestsPath
         const extensionTestsPath = path.resolve(__dirname, './suite/index');
-        console.log('Extension tests path:', extensionTestsPath);
 
         // Download VS Code, unzip it and run the integration test
-        await runTests({ 
-            extensionDevelopmentPath, 
-            extensionTestsPath,
-            launchArgs: [
-                '--disable-extensions',
-                '--verbose',
-                '--extensionDevelopmentPath=' + extensionDevelopmentPath
-            ]
-        });
+        await runTests({ extensionDevelopmentPath, extensionTestsPath });
     } catch (err) {
-        console.error('Failed to run tests', err);
+        console.error('Failed to run tests');
         process.exit(1);
     }
 }